removed unused components, authentication component changed to user-display
