+++ src/types/notificationType.ts
@@ -1,16 +1,16 @@
 type NotificationType = {
   id: number;
   type: string;
-  user: number;
-  requestedEmployee: number;
-  message: string | null;
+  users: string;
+  title: string;
+  message: string;
 };
 
 export type NotificationPostType = {
   type: string;
-  user: number;
-  requestedEmployee: number;
-  message?: string;
+  title: string;
+  users: string;
+  message: string;
 };
 
 export default NotificationType;
